module_upload_fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
aniso8601==9.0.1  # Keep or remove based on usage
attrs==23.2.0  # Keep
blinker==1.8.2  # Usually used for signals, can be removed if not needed
cffi==1.16.0  # Keep, required by some packages
coverage==7.5.1  # Keep, for test coverage
lxml==5.2.2  # Keep, for XML processing
pillow==10.3.0  # Keep, for image processing
pycparser==2.22  # Keep
PyMuPDF==1.24.4  # Keep, for PDF handling
python-docx==1.1.2  # Keep, for DOCX file handling
python-dotenv==1.0.1  # Keep, for environment variables
python-magic==0.4.27  # Keep, for file type detection
pytz==2024.1  # Keep
requests==2.31.0  # Keep, for HTTP requests
six==1.16.0  # Keep
sortedcontainers==2.4.0  # Keep
urllib3==2.2.1  # Keep
wsproto==1.2.0  # Keep, used by WebSocket protocols
psutil